requirements.txt
python-telegram-bot[job-queue]==21.11.1
loguru==0.7.3
python-dotenv==1.0.0

langchain==0.3.20
langchain-community==0.3.19
langchain-huggingface==0.1.2
langchain-gigachat>=0.2.0
langchain-ollama==0.3.4

faiss-cpu==1.10.0
transformers>=4.45.0
sentence-transformers>=3.1.0
huggingface-hub>=0.25.0
tqdm>=4.66.0
accelerate>=0.34.0

httpx>=0.27.0
httpx[http2]>=0.27.0

numpy<2
markdown==3.7
bs4==0.0.2
pandas==2.2.3
torch==2.6.0
ragas>=0.2.0
datasets>=3.0.0